requirements.txt
# core project
ultralytics>=8.4,<8.5
opencv-python>=4.13,<4.14
numpy>=2.4,<2.5
requests>=2.33,<2.34
python-dotenv>=1.2,<1.3

# paddleocr
paddlepaddle==3.2.0
paddleocr>=3.0,<4.0

# easyocr
torch
torchvision
easyocr>=1.7,<2.0

# serial communication
pyserial>=3.5,<3.6


